src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==23.2.0
bcrypt==4.3.0
billiard==4.2.1
boto==2.49.0
boto3==1.38.15
botocore==1.38.15
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
comm==0.2.2
cryptography==44.0.3
datasets
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dill
docker==7.1.0
executing==2.2.0
fabric==3.2.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
gevent==25.5.1
geventhttpclient==2.3.3
google-pasta==0.2.0
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.1
hnswlib==0.8.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
invoke==2.2.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jq==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
multidict==6.4.3
multiprocess
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
onnx==1.18.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pathos==0.3.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
pox==0.3.6
ppft==1.7.7
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-rapidjson==1.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
redis==6.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
s3transfer==0.12.0
safetensors==0.5.3
sagemaker==2.244.0
sagemaker-core==1.0.34
schema==0.7.7
setuptools==80.4.0
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tblib==3.1.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
tritonclient==2.56.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
uvicorn